% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{add_dimred_future}
\alias{add_dimred_future}
\title{Add the future dimensionality reduction based on RNA velocity information, using the existing dimensionality reduction}
\usage{
add_dimred_future(
  dataset,
  dimred = dynwrap::get_dimred(dataset),
  expression = dataset$expression,
  expression_future = dataset$expression_future
)
}
\arguments{
\item{dimred}{The dimensionality reduction}
}
\description{
Add the future dimensionality reduction based on RNA velocity information, using the existing dimensionality reduction
}
