% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{add_velocity}
\alias{add_velocity}
\title{Add velocity to a dynwrap dataset}
\usage{
add_velocity(dataset, spliced = dataset$expression,
  unspliced = dataset$expression_unspliced)
}
\arguments{
\item{dataset}{A dataset created by \code{\link[=wrap_data]{wrap_data()}} or \code{\link[=wrap_expression]{wrap_expression()}}}
}
\description{
Add velocity to a dynwrap dataset
}
