% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{embed_velocity}
\alias{embed_velocity}
\title{Embed the velocity within a given dimensionality reduction}
\usage{
embed_velocity(velocity, dimred)
}
\arguments{
\item{dimred}{The dimensionality reduction}
}
\description{
Embed the velocity within a given dimensionality reduction
}
