% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orient.R
\name{orient_topology_to_velocity}
\alias{orient_topology_to_velocity}
\title{Reorients the edges of the milestone network to the cell's RNA velocity vectors}
\usage{
orient_topology_to_velocity(
  trajectory,
  velocity = trajectory$velocity,
  expression_source = trajectory,
  expression = dynwrap::get_expression(expression_source, "expression"),
  velocity_vector = dynwrap::get_expression(expression_source, "velocity_vector")
)
}
\arguments{
\item{trajectory}{The trajectory as created by \code{\link[dynwrap:infer_trajectory]{infer_trajectory()}} or \code{\link[dynwrap:add_trajectory]{add_trajectory()}}}

\item{velocity}{The velocity object as generated by [get_velocity()].}

\item{expression_source}{The source of expression, can be "counts", "expression", an expression matrix, or another dataset which contains expression}

\item{expression}{The normalised expression values of genes (columns) within cells (rows). This can be both a dense and sparse matrix.}
}
\value{
The trajectory with oriented *milestone_network* and *progressions*
}
\description{
Reorients the edges of the milestone network to the cell's RNA velocity vectors
}
