% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{scvelo}
\alias{scvelo}
\title{Wrapper for the awesome scvelo package}
\description{
"scvelo" allows you to estimate velocity vectors for each cell and gene in a single-cell expression dataset.
}
