% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred.R
\name{add_dimred_future}
\alias{add_dimred_future}
\alias{embed_velocity}
\title{Add the future dimensionality reduction based on RNA velocity information, using the existing dimensionality reduction}
\usage{
add_dimred_future(
  dataset,
  dimred = dynwrap::get_dimred(dataset),
  expression = dataset$expression,
  velocity_vector = dataset$velocity_vector
)

embed_velocity(
  dataset,
  dimred = dynwrap::get_dimred(dataset),
  expression = dataset$expression,
  velocity_vector = dataset$velocity_vector
)
}
\arguments{
\item{dataset}{A dataset created by \code{\link[dynwrap:wrap_data]{wrap_data()}} or \code{\link[dynwrap:wrap_expression]{wrap_expression()}}}

\item{dimred}{Can be
\itemize{
\item A function which will perform the dimensionality reduction, see \code{\link[dyndimred:dimred]{dyndimred::list_dimred_methods()}}
\item A matrix with the dimensionality reduction, with cells in rows and dimensions (\emph{comp_1}, \emph{comp_2}, ...) in columns
}}

\item{expression}{The normalised expression values of genes (columns) within cells (rows). This can be both a dense and sparse matrix.}
}
\description{
Add the future dimensionality reduction based on RNA velocity information, using the existing dimensionality reduction
}
