% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{add_velocity}
\alias{add_velocity}
\title{Add velocity to a dynwrap dataset}
\usage{
add_velocity(
  dataset,
  spliced = dataset$expression,
  unspliced = dataset$expression_unspliced,
  mode = c("stochastic", "deterministic", "dynamical"),
  n_neighbors = 20L,
  velocity = get_velocity(spliced = spliced, unspliced = unspliced, mode = mode,
    n_neighbors = n_neighbors)
)
}
\arguments{
\item{dataset}{A dataset created by \code{\link[dynwrap:wrap_data]{wrap_data()}} or \code{\link[dynwrap:wrap_expression]{wrap_expression()}}}

\item{spliced}{Spliced expression matrix}

\item{unspliced}{Unspliced expression matrix}

\item{n_neighbors}{Number of neighbors to use.}

\item{velocity}{The velocity object as generated by [get_velocity()].}
}
\description{
Add velocity to a dynwrap dataset
}
