% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{get_velocity}
\alias{get_velocity}
\title{Calculate velocity}
\usage{
get_velocity(
  spliced,
  unspliced,
  mode = c("stochastic", "deterministic", "dynamical", "dynamical_residuals"),
  n_neighbors = 20L,
  var_names = "velocity_genes"
)
}
\arguments{
\item{spliced}{Spliced expression matrix}

\item{unspliced}{Unspliced expression matrix}

\item{n_neighbors}{Number of neighbors to use.}

\item{var_names}{Names of variables/genes to use for the fitting. Can be `"velocity_genes"`, `"all"`, or a set of gene names.}
}
\description{
Calculate velocity
}
