% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{get_velocity}
\alias{get_velocity}
\title{Calculate velocity}
\usage{
get_velocity(spliced, unspliced)
}
\arguments{
\item{spliced}{Spliced expression matrix}

\item{unspliced}{Unspliced expression matrix}
}
\description{
Calculate velocity
}
